Added error handling for 400 errors (Bad Request)

+++ server/src/main/java/byu/edu/cs/tweeter/server/service/LogoutServiceImpl.java
@@ -8,11 +8,14 @@ import byu.edu.cs.tweeter.shared.response.LogoutResponse;
 public class LogoutServiceImpl implements LogoutService {
     @Override
     public LogoutResponse logout(LogoutRequest request) {
+        if (request.getAuthToken() == null) {
+            return new LogoutResponse(false, "Bad Request: no authentication given..");
+        }
         getAuthTokenDAO().deleteAuthToken(request.getAuthToken());
         return new LogoutResponse(true);
     }
 
-    AuthTokenDAO getAuthTokenDAO() {
+    public AuthTokenDAO getAuthTokenDAO() {
         return new AuthTokenDAO();
     }
 }
